requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-SocketIO==5.3.6
Werkzeug==3.0.1
SQLAlchemy==2.0.23
Pillow==10.1.0
numpy==1.26.2
opencv-python-headless==4.8.1.78
scikit-learn==1.3.2
python-dotenv==1.0.0
eventlet==0.33.3
pyotp==2.9.0
insightface==0.7.3
onnxruntime==1.16.3
